Added AppSnackbar and SnackbarState
This commit introduces a custom snackbar component (`AppSnackbar`) and a sealed class (`SnackbarState`) to manage different snackbar states (Error, Success).

The `AppSnackbar` is integrated into the main `App.kt` to display snackbars globally. Various screens (Bookmarked, Details, List) are updated to utilize this new snackbar functionality, allowing them to show success or error messages to the user (e.g., when a song is bookmarked/unbookmarked or if a song is not found).